Refactoring - change the location of parser
