--- src/main/java/utils/ParseException.java
@@ -1,8 +1,0 @@
-package utils;
-
-public class ParseException extends Exception{
-
-    public ParseException(String msg) {
-        super("Fail to extract information: " + msg);
-    }
-}
